Fix another windows related bug in ci tests
